requirements.txt
wxPython == 4.1.1
numpy == 1.20.3
pyserial == 3.4